feat: setup typescript & move all files into src/
