+++ .gitignore
@@ -21,3 +21,4 @@ images/demo.gif
 images/synpress.png
 
 examples
+lib/
