+++ src/commands/metamask.js
@@ -201,6 +201,7 @@ const metamask = {
     return true;
   },
   async importWallet(secretWords, password) {
+    console.log({ secretWords, password });
     await playwright.waitAndClick(
       onboardingWelcomePageElements.importWalletButton,
       await playwright.metamaskWindow(),
@@ -209,8 +210,12 @@ const metamask = {
       },
     );
     await module.exports.optOutAnalytics();
+    console.log('after: optOutAnalytics', {
+      secretWords: secretWords.split(' ').entries(),
+    });
     // todo: add support for more secret words (15/18/21/24)
-    for (const [index, word] of secretWords.split(' ').entries()) {
+    for await (const [index, word] of secretWords.split(' ').entries()) {
+      console.log({ index, word });
       await playwright.waitAndType(
         firstTimeFlowImportPageElements.secretWordsInput(index),
         word,
@@ -1163,6 +1168,10 @@ const metamask = {
     await module.exports.getExtensionDetails();
     await playwright.fixBlankPage();
     await playwright.fixCriticalError();
+
+    console.log({
+      secretWordsOrPrivateKey,
+    });
     if (
       await playwright
         .metamaskWindow()
