+++ src/gen-commands.js
