+++ src/lint.js
