+++ src/pages/metamask/main-page.js
