+++ src/pages/metamask/notification-page.js
